uptime: add support for OpenBSD using utmp
OpenBSD uses the original utmp file format when the other OS use utmpx
format to get infos for boot time and logged users.

By using the `utmp-classic` crate, we can process the logged users
on system from the /var/run/utmp file.
