--- src/uu/uptime/src/platform/mod.rs
@@ -1,14 +1,0 @@
-// This file is part of the uutils coreutils package.
-//
-// For the full copyright and license information, please view the LICENSE
-// file that was distributed with this source code.
-
-#[cfg(not(target_os = "openbsd"))]
-mod unix;
-#[cfg(not(target_os = "openbsd"))]
-pub use self::unix::*;
-
-#[cfg(target_os = "openbsd")]
-mod openbsd;
-#[cfg(target_os = "openbsd")]
-pub use self::openbsd::*;
